pdf-downloader-backend/requirements.txt
Flask==3.1.0
Flask-CORS==5.0.0
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
playwright==1.54.0
reportlab==4.4.3
pillow==11.3.0

